UnesparProjetoLattes/requirements.txt
Django==3.1.5
django-filter==2.4.0
fontawesome-free==5.15.3